Adds support for saving and loading measurements
This commit ports the old in-memory data model to a data model using the CoreData framework to store entities. That way it is possible to keep measurements during app restarts. For this purpose we extended the DataCapturing framework by a CoreData data model. This is not an easy task, since loading a data model from a framework is not as straight forward as from a stand alone app. Luckily there is a question on StackOverflow with an appropriate solution (CAREFUL: The second answer is the important one!) https://stackoverflow.com/questions/42553749/core-data-failed-to-load-model.