Changes the DataCapturingListener to a closure
Getting notified of events from the DataCapturingListener was handled by a listener up until now. The reason was, that the code has been ported from Android. To be more Swift like, this has been changed to use a closure. The commit also adds code to acutal notify the provided handler of new geo locations and successful uploads.