Fixes transmission code to Cyface server
Several issues with the Cyface server API have been fixed. Most importantly registering a new device and transmitting data via Alamofire. The first did react to the wrong HTTP Status code and did not terminate in error cases. The second provided headers that are not valid for the Cyface server and had an issue with the Alamofire session manager being killed to quickly.