Adds data migration from old data model to new
Since we switched to RadVerS version we require a context for each measurement. In RadVerS this is work, leisure and shopping. In Cyface it is bicycle, car and motorbike. This was added to the data model without any data migration logic which is bad for the STADTRADELN app. This commit adds the data migration from the previous old data model without measurement context to the new one with the context.