Add handling 409 responses like 201
A 409 (conflict) response from the server was handled as an error until now. This leads to more and more requests building upon, since measurements producing a 409 are never going to leave the system. Therefore this request changes the behaviour to clean measurements causing a 409 the same way as if the upload was successful.