[STAD-258] Fix wrong timestamps on sensor data (#119)
* Add test for capturing sensor data

* Make data points depend on current time again

We had changed this to kernel boot time, which does not acutally work,
since the time reported via the sensor system is not continuous but
broken up in between by device sleeps.

* Remove unusable test

* Add missing documentation

* Add check to filter late sensor values

The removed test case was not properly deleted from the project file.
This commit achieves this.

* Fix intendation in SensorCapturer

* Add missing remark to SensorCapturer documentation

* Fix error handling of sensor event handlers

Previously the error handling would always fail and even in cases of
errors not leave the handler. This can lead to crashes if an error
occurs. Now errors should lead to the app gracefully ignoring that
sensor value.

* Add explanation why Date() is used for timestamp calculation