[STAD-342] Fix for Uploads with new Model Classes (#135)
* Remove CoreData managed objects from non context code

This commit removes all the CoreData managed objects from usage outside
of persistence layer. All the calls should be savely wrapped in
performAndWait.
This must still be included with the example application.

* Streamline managed object class names

* Rename prevLocationUpdateTimeInMillis to prevLocationUpdateTime

since it is of type Date now

* Add additional error output to DataCapturingService

* Make sure locations are only captured in correct order

* Make LocationCacheEntry implement CustomStringConvertible

* Fix bug causing location to be added twice to tracks

* Add error thrown if a measurement was not in the database

* Update tests with changed persistence code

Everything should compile again. Wether it runs need to be figured out,
as soon as the app has been fixed.

* Fix loadEvents

Fixes the loadEvents method in the persistence layer. To achieve this
several compile errors had to be fixed before. This enable the whole
test suite to work under the new persistence code (although it will
probably not pass yet).

* Fix first and last location never from empty tracks

* Fixed ServerConnectionTest to run

Data transimission is not possible yet, due to changes in XCode 13

* Fix DataCapturingTest to show save problem

It seems measurement status is not correctly saved during stop events.
This commit extends the DataCapturingServiceTest to actually show the
problem.

* Update data model to properly store attributes

* Fix stopping measurement

* Fix saving geo locations to measurements

* Fix Test Suite

* Add missing documentation