[CIA-66] new protocol and format (#136)
* Change binary format to use Protobuf

* Update documentation on thrown errors

* Clean up unused code from old Cyface binary format

* Add correctly unpacking acceleration values from their serialised format

* Remove event serialization test

This is not carried out by hand any more and thus requires no test. Testing successful serialisation of events is tested together with serialisation of whole measurements.

* Add upload code from example app

This code is not working yet. It is mostly copied over from the example
app and extended by TODOs. It must be integrated into the app to receive
the actual data, which will be done in the next commit.

* Support new resumable upload protocol

* Integrate Cyface V3 Protocol into SDK

* Add loading of measurement data to upload

Adds loading of measurement data in the new Protobuf format to the
Upload class.

* Fix app authentication

The authenticator was not initialized properly anymore, after the last
update

* Assemble URL for status request correctly

Previously it just appended the whole URL to the existing URL again.
This should be fixed now, by using the provided URL directly.

* Add missing documentation

* Fix status code on upload request

Previously the client expected a status code of 200 to be returned by a
successful upload request. The actual code is 201, which is correct and
now used.

* Add integration test for the whole upload process

* Fix prepending data format version to binary

* Rebase on current main and integration of proto as Pod

* Fix test suite

* Fix Example Application

* Add missing documentation

* Add updated documentation

* Update link to documentation in README.md

* Remove deprecated description about documentation from README

* Make deserialization code use correct constant

* Add TODO for removing count parameter