[DAT-1089] Registration via app (#143)
* Add new UI project based on SwiftUI

* Add generated project files after rebase

* Integrate DataCapturing in SwiftUI App

This is the first rudimentary integration of the DataCapturingService
into the new SwiftUI App. It does not yet work, but it also starts,
pauses and stopps without crashing.

* Add explanations for background location usage

Without them using location service is not possible

* Handle DataCapturingEvent on main thread

This updates the view model and thus needs to run on the main thread

* Refactor details view during capturing

This makes that view align nicer and show actual time instead of just a
number representing seconds.

* Add data synchronization to SwiftUI app

It does not work yet but it builds and starts

* Complete basic features of SwiftUI App

This is basically an MVP of the SwiftUI App. There are still some
missing features from the original app, but it should be possible to
capture some data and upload it to a Cyface Server.

* Fix measurements not correctly deleted after upload

* Improve formatting of values shown for current measurement

* Update display of distance in measurement overview

* Remove debug code

* Add registration

This currently has still some bugs, but basically everything is
finished.

* Add correct default registration URL

* Fix upload of broken measurementes and add documentation

* Prepare code for PR

* Fix Test Suite
* Add missing documentation

* Fix errors caused by XCodes weird reference system

Some files were not at the correct location and others used wrong
references or were missing. This fix should the old example app make
compile again.