Refactor DataCapturing to use Combine
This large commit does a complete refactoring of the old DataCapturingService. It is now called Measurement. The idea is that there is no central DataCapturingService, but that a new Measurement is created on each measurement. Data processing is handled using the Combine framework now. data storage is completely removed from the DataCapturingService/Measurement. This should reduce the amount of CoreData related crashes. During a measurement data is now only stored and never read.

Business logic that was applied during an active measurement has been removed. It is the responsibility of the calling application to realize such features. The Cyface SDK from now on will only report the values it has captured.