[RFR-885] Integrate Cyface SDK with RFR App (#160)
* Add view model for the control bar view

* Implement working login

* Add permissions for location capturing and data synchronization

* Enable measurement start

* Simplify DataCapturingViewModel

Remove handler for events which is not required here and verbose checks
for self on main thread.

* Remove headings from individual views

These where only clutter. Removal cleans up the interface pretty nicely.

* Update map to show current location

If no GPS is available the map shows the project area in northern saxony
and around the town of Köthen.

* Connect running measurement to UI

* Remove username and password from DataCapturingViewModel

DataCapturingViewModel should only handle the actual data capturing.
Data synchronization should be separate.
Thus, those properties are removed.

* Add height calculation to RFR App

* Connect statistics overview page to SDK

* Add statistics and account registration

* Fix Registration with Cyface Server

* Publish registration successful on main thread

* Add registration template to registration request

* Add Input Validation for RegistrationView

* Refactor to use Combine for Statistics

* Refactor DataCapturing to use Combine

This large commit does a complete refactoring of the old DataCapturingService. It is now called Measurement. The idea is that there is no central DataCapturingService, but that a new Measurement is created on each measurement. Data processing is handled using the Combine framework now. data storage is completely removed from the DataCapturingService/Measurement. This should reduce the amount of CoreData related crashes. During a measurement data is now only stored and never read.

Business logic that was applied during an active measurement has been removed. It is the responsibility of the calling application to realize such features. The Cyface SDK from now on will only report the values it has captured.

* Refactored to accomodate current database infrastructure.

* Add raw sensor data to knew workflow

* Add Bike Rally information

* Show track in Measurement View

* Add Voucher View for users registered with for the community of Schkeuditz

* Remove obsolete code lines.

* Add Support for resuming after long pause.

* Add Privacy Policy to App Start

* Release Beta 0.2.2

* Change login to using OAuth2 / OpenID

* Fix bug in track length calculation of FinishedMeasurement

* Fix Privacy Policy View

* Fix warning in VoucherViewModel

* Update Login to use AppAuth Framework

* Add Upload with AppAuth

* Remove Privacy Policy from App

This is now provided via the embedded OAuth View.

* Change architecture to make new measurements appear after stop

* Make measurements not scroll over top toolbar

* Add button to restart authentication process

* Prepare for Release

Adding a few logging statements and using the production URL

* Remove Alamofire

Since it does not support background uploads there is no need to keep that dependency around. So it gets removed and replaced by Apples native URLSession API.

* Add missing License headers

* Add missing documentation to RFRError

* Add missing linebreaks to RFRError

* Add Delete User Functionality to RFR App

* Add test case to diagnose upload problems

* Separate different build environments

Makes three build environments with three configurations for local server, staging server and production server.

* Add storing upload state

This was missing and thus lost after each app restart.

* Integrate Apple Review Feedback

* Remove Bike Ralley Markers from Main Map

* Add log output to diagnose upload problems

* Release Version 3.0.0

* Provide RFR App Localization

* Fix wrong translation of "Daten übertragen"

* Display RFR Logo correctly

* Filter live speed values below zero.

* Fix measurement not starting if permissions dialog is shown.

* Add correct endpoint for collector to Production Setup

* Update Formula to calculate avoided CO2 emissions

* Cleanup code

Add documentation and cleanup some dead code.