Actually use the finish handler of the CapturedDataStorage
No finish event as implemented in the last commit is necessary, since our API has an explicit handler on the CapturedDataStorage, that is called as soon as data storage has finished. This should be used to store everything to the database and carry out clean up code at the end of a measurement.